refactor: restructure dynamic content initializer
